+++ src/main/java/org/openlmis/requisition/exception/JaversExistingEntryException.java
@@ -13,25 +13,16 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
  */
 
-package org.openlmis.requisition.repository;
+package org.openlmis.requisition.exception;
 
-import java.io.Serializable;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.NoRepositoryBean;
-import org.springframework.data.repository.PagingAndSortingRepository;
+import org.openlmis.requisition.utils.Message;
 
 /**
- * Extension of {@link PagingAndSortingRepository} to enable using generic parameters
- * in creating Javers logs and provide additional methods to retrieve entities
- * using the pagination and sorting abstraction to ensure.
+ * Signals an existing entry when trying to add another snapshot by Javers.
  */
-@NoRepositoryBean
-public interface BaseAuditableRepository<T, I extends Serializable>
-    extends PagingAndSortingRepository<T, I> {
+public class JaversExistingEntryException extends BaseMessageException {
 
-  /**
-   * Returns a {@link Page} of entities which there are no Javers logs created for.
-   */
-  Page<T> findAllWithoutSnapshots(Pageable pageable);
+  public JaversExistingEntryException(Message message) {
+    super(message);
+  }
 }